requirements/base.txt
djangorestframework>=3.14.0
shapely>=1.8.4
scikit-learn>=1.1.2
geohash2>=1.1
